etl/requirements.txt
google-api-python-client==2.151.0
google-auth==2.35.0
supabase==2.9.0
